Add files via upload
This tool will track lines in the red (638 nm) and blue (488 nm) channels, and determine the binding lifetimes and orders of assembly and disassembly of the events. The two channels of colocalization can be changed by selecting all, and replacing all the instances of a color with the new color. That is, to compare green and red replace all instances of "blue" with "green". Note each channel has to first be tracked individually before performing the colocalization analysis.

At the end of this script the event lifetimes, binding positions, colocalization categories, and duration of colocalizations will output as .txt files.